+++ src/renderer/src/components/OverworldMode.tsx
@@ -1,30 +1,54 @@
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import React, { ReactElement, SyntheticBaseEvent, useState } from "react";
 
-import { MapObjectArgs, OverworldLayerType, OverworldModeProps } from '../../../common/types';
 import { getOverworldTypeFactory } from '../../../common/objects';
+import { MapObjectArgs, OverworldLayerType, OverworldModeProps } from '../../../common/types';
 import OverworldGridCanvas from './Overworld/OverworldGridCanvas';
+import OverworldLayerControls from './Overworld/OverworldLayerControls';
+import OverworldObjectControls from './Overworld/OverworldObjectControls';
 import OverworldObjectOptions from './Overworld/OverworldObjectOptions';
 
 function OverworldMode( props: OverworldModeProps ): ReactElement {
 	const { exitMode, graphics, overworld, palettes, setOverworld } = props;
+	const [ selectedLayer, setSelectedLayer ] = useState<number>( 0 );
 	const [ selectedObject, setSelectedObject ] = useState<number | null>( null );
 	const [ selectedObjectType, setSelectedObjectType ] = useState<number>( 0 );
 
+	const layers = overworld.getLayersList();
 	const typesFactory = getOverworldTypeFactory( OverworldLayerType.block );
 
+	const addLayer = (): void => setOverworld( overworld.addLayer( OverworldLayerType.block ) );
+
+	const moveLayerDown = (): void => {
+		if ( selectedLayer >= layers.length - 1 ) {
+			return;
+		}
+		setOverworld( overworld.moveLayerDown( selectedLayer ) );
+		setSelectedLayer( selectedLayer + 1 );
+	};
+
+	const moveLayerUp = (): void => {
+		if ( selectedLayer <= 0 ) {
+			return;
+		}
+		setOverworld( overworld.moveLayerUp( selectedLayer ) );
+		setSelectedLayer( selectedLayer - 1 );
+	};
+
+	const removeLayer = (): void => {
+		setOverworld( overworld.removeLayer( selectedLayer ) );
+		setSelectedLayer( Math.max( 0, selectedLayer - 1 ) );
+		setSelectedObject( null );
+		setSelectedObjectType( 0 );
+	};
+
 	const removeObject = (): void => {
-		setOverworld( overworld.removeObject( selectedObject ) );
+		setOverworld( layers[ selectedLayer ].removeObject( selectedObject ) );
 		setSelectedObject( null );
 	};
 
 	const updateObject = ( index: number, object: MapObjectArgs ): void => {
-		setOverworld( overworld.updateObject( index, object ) );
-	};
-
-	const updateSelectedObjectType = ( e: SyntheticBaseEvent ): void => {
-		const value = e.target.value;
-		setSelectedObjectType( value );
+		setOverworld( layers[ selectedLayer ].updateObject( index, object ) );
 	};
 
 	return <div>
@@ -33,22 +57,31 @@ function OverworldMode( props: OverworldModeProps ): ReactElement {
 			graphics={ graphics }
 			overworld={ overworld }
 			palettes={ palettes }
+			selectedLayer={ selectedLayer }
 			selectedObject={ selectedObject }
 			selectedObjectType={ selectedObjectType }
 			setOverworld={ setOverworld }
 			setSelectedObject={ setSelectedObject }
 		/>
-		<div>
-			<label>
-				<h2>Object type to add:</h2>
-				<select value={ selectedObjectType } onChange={ updateSelectedObjectType }>
-					{ typesFactory.map( ( type, i ) => <option key={ i } value={ i }>{ type.name }</option> ) }
-				</select>
-			</label>
-		</div>
+		<OverworldLayerControls
+			addLayer={ addLayer }
+			layers={ layers }
+			moveLayerDown={ moveLayerDown }
+			moveLayerUp={ moveLayerUp }
+			removeLayer={ removeLayer }
+			selectedLayer={ selectedLayer }
+			setSelectedLayer={ setSelectedLayer }
+			setSelectedObject={ setSelectedObject }
+			setSelectedObjectType={ setSelectedObjectType }
+		/>
+		<OverworldObjectControls
+			typesFactory={ typesFactory }
+			selectedObjectType={ selectedObjectType }
+			setSelectedObjectType={ setSelectedObjectType }
+		/>
 		{ selectedObject !== null && <OverworldObjectOptions
 			removeObject={ removeObject }
-			selectedObject={ overworld.getObject( selectedObject ) }
+			selectedObject={ layers[ selectedLayer ].getObject( selectedObject ) }
 			selectedObjectIndex={ selectedObject }
 			typesFactory={ typesFactory }
 			updateObject={ updateObject }
